Create horizontal display for hourly cards move cleaner functions to separate file
