+++ src/App.js
@@ -5,6 +5,7 @@ import HourlyCard from './Hourly-Card.js'
 import sun from './sun.png'
 import data from './mock-data.js';
 import HourlyContainer from './Hourly-Container.js';
+import cleaner from './data-cleaner.js'
 
 class App extends Component {
   constructor() {
@@ -13,46 +14,18 @@ class App extends Component {
       hourlyCards: [...data.hourly_forecast],
       hourlyCardLimit: 7
     }
-    this.fillHourlyCards = this.fillHourlyCards.bind(this);
+    cleaner.fillHourlyCards = cleaner.fillHourlyCards.bind(this);
   }
 
-  fillHourlyCards() {
-  let hoursObj = this.state.hourlyCards.reduce((hoursprojection, hour) => {
-    if(hoursprojection.length > 7) {
-        return hoursprojection;
-      }
-      hoursprojection.push({hour: hour.FCTTIME.hour , 
-                            condition: hour.condition ,
-                            icon: hour.icon_url, 
-                            temp: hour.temp.english})
-      return hoursprojection;
-    }, [])
-    return hoursObj;
-}
-
-  getCurrentData(data) {
-    const weatherStatus = {
-    location: data.current_observation.display_location.full,
-    currentCondition: data.forecast.simpleforecast.forecastday[0].conditions,
-    day: data.forecast.simpleforecast.forecastday[0].date.weekday,
-    temp: data.current_observation.temp_f,
-    high: data.forecast.simpleforecast.forecastday[0].high.fahrenheit,
-    low: data.forecast.simpleforecast.forecastday[0].low.fahrenheit,
-    summary: data.forecast.txt_forecast.forecastday[0].fcttext,
-    icon: data.forecast.simpleforecast.forecastday[0].icon_url
-  }
-
-    return weatherStatus;
-  }
 
   render() {
-    let {location, currentCondition, summary, temp, high, low, icon} = this.getCurrentData(data);
+    let {location, currentCondition, summary, temp, high, low, icon} = cleaner.getCurrentData(data);
     return (
       <div className="App">
-        <Card name={location} condition={icon} description="A lovely sun" currentTemp={temp} highLow={high, low} summary={summary}/>
+        <Card city={location} condition={icon} description="A lovely sun" currentTemp={temp} high={high} low={low} summary={summary}/>
         <HourlyContainer>
         {
-          this.fillHourlyCards().map(hour => {
+          cleaner.fillHourlyCards().map(hour => {
             return <HourlyCard hour={hour.hour} condition={hour.condition} projectedTemp={hour.temp} icon={hour.icon} />
           }) 
         }
