+++ src/Card.js
@@ -12,7 +12,7 @@ class Card extends Component {
         <h2>{this.props.city}</h2>
         <h3><img src={this.props.condition} alt={this.props.description} /></h3>
         <h3>{this.props.currentTemp}</h3>
-        <h3>{this.props.highLow}</h3>
+        <h3 className="highLow">{this.props.high}</h3>/<h3 className="highLow">{this.props.low}</h3>
         <p>{this.props.summary}</p>
       </div>
     )
